Dashboard cards config (#749)
