+++ .changeset/lazy-goats-perform.md
@@ -1,0 +1,5 @@
+---
+'@keystatic/core': patch
+---
+
+Observe nav config as dashboard cards.
